feat!: Update for OpenBSD 7.7
This commit enables the ports to build on OpenBSD 7.7 but at the expense of earlier OpenBSD versions.

1. OpenBSD 7.7 ports infra for python was refactored to use different variables for package versions which trickles down to how the tarballs are retrieved from pypi. These changes are incompatible with OpenBSD ports < 7.7.
2. The cmake port in OpenBSD 7.7 was refactored into a meta and multiple sub ports, causing a change in dependency declaration for py-awscrt port.

For these reasons, with this commit, the ports will not build on earlier OpenBSD versions.

Additionally, cherry pick a change from AWS CLI to remove requirements for zipp. This upstream change works in our favor since the version of zipp in ports doesn't meet the (previously set) version requirements